stablized api
Implementation is completely revised with the changes in the approach of `ImpVec:1.0.0`. In particular, unsafe code is limited to a few methods which are safely isolated and documented. This crate does not use any unsafe code other than these imp-vec methods specialized for building self referential collections.

Further, the implementation makes use of the updates in the `PinnedVec` implementation. In particular, the `Node` composes `SelfRefNext` and `SelfRefPrev` trait to define its relations with the other nodes in the vector. This allows to directly use `ImpVec`s `set_next` and `set_prev` methods; note that these methods are safe.

Recursive nature of the linked list data structure is emphasized. Note that each node itself when considered as a front or back must be able to represent a linked list. This flexibility is provided through slices. It is now cheap and very flexible to recursively define linked lists out of linked lists by slicing. These slices are immutable; however, can directly be collected into owned lists whenever required.

* `LinkedListSlice`s are defined together with convenient, flexible and cheap methods to create them recursively.
* `FromIterator` is implemented.
* `Extend` is implemented.
* `Default` is implemented.
* Unsafe code is limited to specialized calls to a couple of `ImpVec` methods.
* `PinnedVec`s composable self referential element traits are made use of.
* Constructors are simplified.
* Benchmarks added.
* Documentation revised.
* Tests are revised and coverage is improved.