list made generic over pinned vec storage
* `List<V, M> is changed as `List<V, M, P>` where
  * `V` is the variant,
  * `M` is the memory policy,
  * `P` is the pinned vec used to store the nodes.

This is not a breaking change since the additional type parameter has a default value `P = SplitVec<Node<V>, Recursive>`. This default is great for the linked list and we really do not need to use any other pinned vec variant.

However, specialized data structures wrapping the linked list might benefit from this flexibility. An immediate possible data structure is the `Tour` which is composed of an indices vec and a permutation stored as a linked list. In most cases, tour has a fixed length or a well-known upper bound on length. Then, we can benefit from using a `FixedVec` rather than a `SplitVec`. This was not possible in the previous version; however, with the additional generic type argument for the pinned vec, it will be possible.