list made generic over pinned vectors
List made generic over pinned vectors. This is important as it allows certain data structures which has known size to use a `FixedVec` rather than a `SplitVec`. This is beneficial especially in small collections.

However, default type is set as the split-vec with recursive growth strategy. Therefore, although the type definition is extend by an additional generic parameter, the changes do not break the api.

Furthermore, `pointers` iterators are provided both for singly and doubly lists.

Equality implementations are extended for convenience.